+++ tests/notification/test_sites.py
@@ -22,7 +22,7 @@ class TempSiteID():
         self.site_id_temp = site_id
 
     def __enter__(self):
-        self.site_id_stable = getattr(settings, 'SITE_ID', 1)
+        self.site_id_stable = getattr(settings, 'SITE_ID', None)
         settings.SITE_ID = self.site_id_temp
 
     def __exit__(self, *args):
@@ -31,39 +31,56 @@ class TempSiteID():
 
 class SiteIDTest(TestCase):
 
+    def setUp(self):
+        super(SiteIDTest, self).setUp()
+        self.second_site = Site.objects.create(domain='generic_second_site')
+
+    @override_settings()
     def test_on_site_respects_SITE_ID_setting(self):
+        site_1 = Site.objects.get(id=1)
+        site_2 = Site.objects.create()
         site_1_pages = 3
         site_2_pages = site_1_pages + 1
-        site_2 = Site.objects.create()
+
         index = 0
         for i in range(site_1_pages):
-            EmailNotification.objects.create(name=index)
+            notification = EmailNotification.objects.create(name=index)
+            notification.sites.add(site_1)
             index += 1
         for i in range(site_2_pages):
             notification = EmailNotification.objects.create(name=index)
-            notification.sites.add(site_2)  # site_1 is already added
+            notification.sites.add(site_2)
             index += 1
 
-        self.assertEqual(EmailNotification.objects.on_site().count(), site_1_pages + site_2_pages)
+        with TempSiteID(site_1.id):
+            self.assertEqual(EmailNotification.objects.on_site().count(), site_1_pages)
+
         with TempSiteID(site_2.id):
             self.assertEqual(EmailNotification.objects.on_site().count(), site_2_pages)
 
-    def test_site_not_added_multiple_times_on_save(self):
+    @override_settings()
+    def test_site_not_overriden_on_save(self):
         site = Site.objects.create()
         # site_id will be a string on live, since its an environment variable
         site_id = str(site.id)
         with TempSiteID(site_id):
             email = EmailNotification.objects.create(name='test_name')
-            email.save()
+            email.sites.add(2)
             email.save()
         self.assertEqual(email.sites.count(), 1)
+        self.assertNotEqual(email.sites.first().id, site_id)
+        self.assertEqual(email.sites.first().id, self.second_site.id)
 
+    @override_settings()
     def test_multiple_added_sites_are_reflected_by_on_site(self):
         site_2 = Site.objects.create()
         notification = EmailNotification.objects.create()
+        notification.sites.add(1)
         notification.sites.add(site_2)
 
-        self.assertIn(notification, EmailNotification.objects.on_site())
+        with TempSiteID(1):
+            self.assertIn(notification, EmailNotification.objects.on_site())
+
         with TempSiteID(site_2.id):
             self.assertIn(notification, EmailNotification.objects.on_site())
 
@@ -72,6 +89,9 @@ class SiteRequestTest(TestCase):
 
     def setUp(self):
         super(SiteRequestTest, self).setUp()
+        self.site = Site.objects.get(id=1)
+        self.site.domain = 'testserver'
+        self.site.save()
         User.objects.create_user(username='dummy', password='dummy')
         self.client.login(username='dummy', password='dummy')
         user = User.objects.get(username='dummy')
@@ -81,21 +101,12 @@ class SiteRequestTest(TestCase):
         self.report.save()
         self.submit_url = reverse('test_submit_report', args=[self.report.pk])
 
-    @override_settings()
-    @patch('django.http.request.HttpRequest.get_host')
-    def test_can_request_pages_without_site_id_set(self, mock_get_host):
-        mock_get_host.return_value = Site.objects.get(id=settings.SITE_ID).domain
-        del settings.SITE_ID
+    def test_can_request_pages_without_site_id_set(self):
         response = self.client.get(self.submit_url)
         self.assertNotEqual(response.status_code, 404)
 
-    @override_settings()
-    @patch('django.http.request.HttpRequest.get_host')
     @patch('callisto.notification.managers.EmailNotificationQuerySet.on_site')
-    def test_site_passed_to_email_notification_manager(self, mock_on_site, mock_get_host):
-        mock_get_host.return_value = Site.objects.get(id=settings.SITE_ID).domain
-        site_id = settings.SITE_ID
-        del settings.SITE_ID
+    def test_site_passed_to_email_notification_manager(self, mock_on_site):
         self.client.post(
             self.submit_url,
             data={
@@ -106,4 +117,4 @@ class SiteRequestTest(TestCase):
                 'key': self.report_key,
             },
         )
-        mock_on_site.assert_called_with(site_id)
+        mock_on_site.assert_called_with(self.site.id)